+++ KU.sql:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=http://kusmartfarm.synology.me/phpmyadmin/export.php
